Convert the module to a class, implement promises, implement limit, and moved to the index file
Sorry this couldn't be more atomic. This is a big one... The module is now an es6 class. Instead of using callbacks, promises are implemented, which will fix #64. Limit.js is also used which was talked about in PR #74. Instead of using the index file to just point to the class in the lib, move the class into the index file. It seemed a little redundant to me.