Include default id parameter when none is present
Some endpoints (like watchlist endpoints) have an optional id parameter. So by default, if there's a session id and the endpoint has an id parameter and none was passed to the function, it will use the default {account_id} as the id parameter. Cleaned up the tests since they all called the same assertions. And added a test for this default id paramter.